feat: Add API for building attribute getter free functions. (#1704)
Adds a new RendererServices API build_attribute_getter which allows for building custom free functions that provide values for attributes, effectively replacing both get_attribute and get_array_attribute.

This new API is invoked at material compilation time to allow render developers to provide specialized functions capable of taking advantage of known compile time information. This will allow for many optimization opportunities, such as replacing run-time map lookups with direct memory reads from known locations, remove brancing required for checking if derivatives are needed, branching on type-conversions if required, etc.

This PR currently implements this approach for attributes, a future PR could do the same for user-data.
Tests

TestShade provides an example implementation showing how this compile time information can be used to select an appropriate function to use as the attribute provider, and how to configure the function signature.

---------

+++ src/liboslexec/backendllvm.h
@@ -152,7 +152,7 @@ public:
 
     llvm::Value* llvm_load_stringhash(ustring str)
     {
-        return ll.constant64(str.hash());
+        return ll.constant64((uint64_t)str.hash());
     }
 
     /// Legacy version
